Hack to fix cutworm and cleaver
Problem: When the cleaver cleaves an adjacent target, cutworm() uses
u.dx and u.dy to determine which segment should be cut, which needs to
be adjusted to the actual direction the adjacent target is from the
player. Otherwise one hits the impossible() in cutworm because it
checks the primary attack location for adjacent attacks.